+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog - Weather plugin for glFusion
 
+## Version 1.0.4 -
+- Add apixu.com weather provider
+- Add changes from matrox66 for World Weather Online class
+- Extract common weather functions into base_api.class.php
+
 ## Version 1.0.3 - 2016-11-06
 - Hide large description from display on small viewports
 - Remove World Weather Online as preferred provider